Change application directory structure!!
- Setting "Absolute Path".
- Directory consist of "preview", "setting" and "export".
    - "preview": Domain of banner preview.
    - "setting": Domain of setting banner components.
    - "export": Domain of export with download or clipboard.

* TODO
    * Fixed path bug and issue!
