+++ babel.config.js
@@ -12,11 +12,9 @@ module.exports = {
   ],
   plugins: [
     [
-      'module-resolver', {
-        root: ['./src'],
-        alias: {
-          '~/*': './src',
-        },
+      'babel-plugin-root-import', {
+        rootPathPrefix: '~',
+        rootPathSuffix: 'src',
       },
     ],
   ],
