+++ src/domain/BannerMakerHeader.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-const AppHeader = () => (
+const BannerMakerHeader = () => (
   <>
     <h1>Banner+</h1>
     <h5>⚙️ Settings</h5>
   </>
 );
 
-export default AppHeader;
+export default BannerMakerHeader;
