+++ src/domain/preview/ImageLayerContainer.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useSelector } from 'react-redux';
 
-import { MemoizedImageLayer } from './ImageLayer';
+import { MemoizedImageLayer } from './components/ImageLayer';
 
 const ImageLayerContainer = ({ onDraw }) => {
   const backgroundImage = useSelector((state) => state.backgroundImage);
