+++ src/domain/preview/components/BackgroundLayer.test.jsx
