+++ src/domain/preview/components/ImageLayer.test.jsx
