+++ src/domain/setting/GeneralContainer.jsx
@@ -2,17 +2,17 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import BannerSizeButton from './BannerSizeButton';
-
-import Card from './component/Card';
-import CardTitle from './component/CardTitle';
-import CardSubTitle from './component/CardSubTitle';
+import Card from 'components/card/Card';
+import CardTitle from 'components/card/CardTitle';
+import CardSubTitle from 'components/card/CardSubTitle';
 
 import {
   setWidth,
   setHeight,
   setWidthHeight,
-} from './slice';
+} from 'slice';
+
+import BannerSizeButton from './components/BannerSizeButton';
 
 const BackgroundContainer = () => {
   const dispatch = useDispatch();
