+++ src/domain/setting/components/FontSize.test.jsx
